+++ client/App.tsx
@@ -1,5 +1,5 @@
 import "./global.css";
-
+import { HelmetProvider } from "react-helmet-async";
 import { Toaster } from "@/components/ui/toaster";
 import { createRoot } from "react-dom/client";
 import { Toaster as Sonner } from "@/components/ui/sonner";
@@ -53,6 +53,7 @@ const App = () => {
         <BrowserRouter>
           <CartProvider>
             <StoryProvider>
+            <HelmetProvider>
               <Routes>
                 <Route path="/" element={<Index />} />
                 <Route path="/sell" element={<Sell />} />
@@ -75,6 +76,7 @@ const App = () => {
                 {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
                 <Route path="*" element={<NotFound />} />
               </Routes>
+              </HelmetProvider>
             </StoryProvider>
           </CartProvider>
         </BrowserRouter>